requirements.txt
Flask~=3.1.1
django-model-utils~=5.0.0
streamlit~=1.45.1
python-dotenv~=1.1.0
numpy~=2.2.6
pandas~=2.2.3
scikit-learn~=1.6.1